# openapi-user.yaml
openapi: 3.0.0
info:
  title: MCCS Alpha User API
  description: |
    ## Introduction

    The MCCS Alpha User API v1 exposes all of the end user functionality currently available in the [alpha version of the MCCS API](https://github.com/technoshantoms/mccs-alpha-api).

    By providing an API, developers who want to create their own front-end user interface for MCCS will have significant flexibility to implement it in whatever way they choose. This means developers can present MCCS in any language, setup their own signup flow, optimize it for whatever devices their users prefer, develop a mobile app, integrate other services such as chat, etc.

    Importantly, an API enables developers to integrate MCCS functionality directly into their own apps (e.g., import transfer data into an accounting application, instruct mutual credit transfers from an e-wallet application, interact with MCCS data via a chat bot, etc.).

    ## Background

    To understand how the MCCS Alpha API works, please read the following design documents:

    - [MCCS Alpha Data Model](https://github.com/technoshantoms/mccs/blob/master/alpha-data-model.md)
    - [MCCS Alpha Functionality](https://github.com/technoshantoms/mccs/blob/master/alpha-functionality.md)

    ## Test Server

...

    ## License
  license:
    name: MIT License
    url: https://github.com/technoshantoms/mccs-alpha-api/blob/dev/LICENSE
  version: '1'
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /signup:
    post:
      tags:
        - Manage Account
      summary: Create a new account - a user resource and associated entity resource
      description: |
        Individual users can create an account in MCCS by providing an email address, creating a password and adding some other details about themselves and their "business". A "business" need not be a formally established business; it could simply be a list of their skills that they are willing to offer to other participants in the network, or another type of entity such as an association, not-for-profit, NGO, etc.

        We use the term **entity** to generically identify businesses, non-profits, NGOs, sole proprietorships, limited companies, etc.

        The `POST /signup` API call creates a user resource and an entity (e.g., a business) resource and links the two together. The user resource references its associated entity, and the entity resource references its associated user. These two related resources make up what is referred to as an account.

        An account can be created with only a unique email address and a password, enabling a new user to get started quickly. The `userID` and `entityID` are returned in the response body, along with a [JSON Web Token](https://jwt.io) to authorize the user of the account for further API calls related to the user and entity resources just created.
        
        All other details (entity name, user first/last names, goods/services offered or wanted, etc.) are optional and can be added later. The decision about customer data requirements for signup can be determined by developers implementing a front end to the API, since they will be working with the organizations implementing MCCS. If other information is collected during the initial registration process, these extra details can be passed immediately in this `POST /signup` API endpoint or later using the `PATCH /user` and/or `PATCH /user/entities/{entityID}` API endpoints.
        
        The current (v1) implementation of the API has a one-to-one relationship between the user and the entity. This will be changed to a many-to-many relationship in a future implementation.
      requestBody:
        description: Signup request body
        required: true
        content:
          application/json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/SignupRequiredFields'
              - $ref: '#/components/schemas/SignupAllFields'
            examples:
              requiredFields:
                value:
                  userEmail: jdoe@dev.null
                  password: 1TrulySecurePassword!!
              allFields:
                value:
                  userEmail: jdoe@dev.null
                  password: 1TrulySecurePassword!!
                  firstName: Jane
                  lastName: Doe
                  userPhone: "+442010203040"
                  entityEmail: nwcltd@dev.null
                  entityName: New World Consulting Limited
                  incType: ltd
                  companyNumber: A12345
                  entityPhone: "+442090807060"
                  website: https://neworco.null
                  declaredTurnover: 20000
                  description: "We show you how good things can be and what you need to do to make them happen."
                  address: "123 Yellow Brick Road"
                  city: "London"
                  region: "Greater London"
                  postalCode: "UK1 2ENG"
                  country: "England"
                  showTagsMatchedSinceLastLogin: true
                  receiveDailyMatchNotificationEmail: true
                  offers:
                    - consulting
                    - system-design
                  wants:
                    - it-services
                    - accounting
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userID:
                        type: string
                      entityID:
                        type: string
                      token: 
                        type: string
              example:
                data:
                  userID: "5eec78f4a880b7c235f66e80"
                  entityID: "5eec78f4a880b7c235f66e7c"
                  token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1OTI2NDIxNjQsInVzZXJJRCI6IjVlZWM3OGY0YTg4MGI3YzIzNWY2NmU4MCIsImFkbWluIjpmYWxzZX0.WhPMfV9S-xgOHcuMT_K-fBhR_K6MpXiJ15GYn4Jz7im1dvhwnV2bEnwuWeFUockl45StxguvIOA5qJ-_3xA14CuP0wJbZa3hVH4jnYXonHlCyHDB8w67RLN9IMFGnSEshhh4D3RjQVpEpBm7jLhQcHKOSQIqUU_RfPkiNxpUkDI6t1RW_-rhY4UsTTuxnC5SOeajzOgiDFM4NwJfjebys8xDGTqYoi4dpCJZEtD_U_X9BuEOovRRJo0TY6m76XxUB9J5U_Hfjm7k_A3aLv3WgDScRv_k-LSsOvviGk1A2ct0nQ2RaVY1udA-76rv1xbvSd26Xds2XtrPb_SzUL-J8A"
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
  /login:
    post:
      tags:
        - Manage Account
      summary: Login and receive an authorization token
      description: |
        A user will need to authenticate with an email/password combination which, if successful, will result in a JSON Web Token (JWT) being passed to the user to use with each API request that requires authorization.

        The JWT will expire after 24 hours, or as soon as the user calls the `/logout` API.
      requestBody:
        $ref: '#/components/requestBodies/loginUser'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
              example:
                data:
                  token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1OTI5ODU0MzcsInVzZXJJRCI6IjVlZjFiNWQ5NTA2ODQ3ZWE5MTRkYjA2NiIsImFkbWluIjpmYWxzZX0.E3YJWpUN0h2LS3IfE2niYclTnhKUGXDay7SR_VeVzaq4a_lZPs2w4yxFUWbbeAcPeMgtcd3bEoX6PtDhW3hjJ8XVbqmkJQRmh3uw5ULfgzkIzrQw3twaG8TO6ARM4UvKJWz2Wqb6czd5SesJmM2htIuY3DBJ_u4r9x3hshM5_0kHalfEZtQvae0KrJ2_eBjPmFOdON62QungzStkjTKTfqsvystwFSdfOwAltg0Nri1Z6q-E9AZMTnAxzKNqjp4Ja3hX1IoZXPiV8F0-yl1PhrKI_YzP57pOt84T72_WK3z1_hofHAOwby5-2qvtoKWSmxqzKsIYUYeG89W7h4TvBQ"
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500:
          $ref: '#/components/responses/ServerError'
  /logout:
    post:
      tags:
        - Manage Account
      summary: Logout and expire the authorization token
      description: Logging out will immediately expire the JWT currently associated to the user's account.
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
  /password-reset:
    post:
      tags:
        - Manage Account
      summary: Request a password reset token
      description: When a password is lost by the user, a new one can be requested by providing the email address associated with the user's account. A password reset token will then be sent to that email address.
      requestBody:
        $ref: '#/components/requestBodies/resetPassword'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
  /password-reset/{token}:
    post:
      tags:
        - Manage Account
      summary: Reset a password
      description: The new password can be set by providing it along with the password reset token that was received at the user's email address.
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        $ref: '#/components/requestBodies/updatePassword'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
  /password-change:
    post:
      tags:
        - Manage Account
      summary: Change the password
      description: A logged-in user can change the password by sending the new password along with the JWT.
      requestBody:
        $ref: '#/components/requestBodies/updatePassword'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
  /user:
    get:
      tags:
        - Manage Account
      summary: View a user's own details
      description: Logged-in users can request their details as recorded in the MCCS database.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
              example:
                data:
                  id: 5eec78f4a880b7c235f66e80
                  email: jdoe@dev.null
                  firstName: Jane
                  lastName: Doe
                  telephone: "+442010203040"
                  lastLoginIP: 192.168.12.34
                  lastLoginDate: "2020-06-19T23:59:59.999Z"
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
    patch:
      tags:
        - Manage Account
      summary: Modify a user's own details
      description: |
        Users can change their details in MCCS' database, except for the `email` field which can only be changed by an administrator.
        
        The `id`, `lastLoginIP` and `lastLoginDate` fields are system-generated and therefore are not changeable by either users or administrators.
      requestBody:
        $ref: '#/components/requestBodies/updateUser'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
              example:
                data:
                  id: 5eec78f4a880b7c235f66e80
                  email: jdoe@dev.null
                  firstName: Jane K.
                  lastName: Doer
                  telephone: "+442012345678"
                  lastLoginIP: 192.168.12.34
                  lastLoginDate: "2020-06-19T23:59:59.999Z"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
...
components:
  schemas:
    SignupRequiredFields:
      type: object
      title: SignupRequiredFields
      required:
        - userEmail
        - password
      properties:
        userEmail:
          type: string
        password:
          type: string
...
  parameters:
    token:
      name: token
      description: The password reset token.
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request is missing a required parameter.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
          example:
            errors:
              - message: <named> parameter is missing.
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT